Code/requirements.txt
autopep8==1.5.4
certifi==2020.12.5
click==7.1.2
dnspython==2.0.0
email-validator==1.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycodestyle==2.6.0
python-dotenv==0.15.0
SQLAlchemy==1.3.22
toml==0.10.2
Werkzeug==1.0.1
wincertstore==0.2
WTForms==2.3.3
